requirements.txt
async-timeout==5.0.1
azure-core==1.37.0
azure-cosmos==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
paho-mqtt==2.1.0
redis==7.0.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
pymongo
pytz